fix(tests): update stale imports to renamed utility modules
Update test files that still imported from deleted modules after the
date-utils consolidation and repoAnalytics rename.
